+++ lib/src/ui/views/steps/customize/customize_view.dart
@@ -4,6 +4,7 @@ import 'package:flutter_installer/src/ui/global/app_colors.dart';
 import 'package:flutter_installer/src/ui/global/ui_helpers.dart';
 import 'package:flutter_installer/src/ui/widgets/custom_button.dart';
 import 'package:flutter_installer/src/ui/widgets/expanded_container.dart';
+import 'package:flutter_installer/src/ui/widgets/text_link.dart';
 
 import 'package:stacked/stacked.dart';
 
@@ -20,6 +21,7 @@ class CustomizeView extends StatelessWidget {
   Widget build(BuildContext context) {
     return ViewModelBuilder<CustomizeViewModel>.reactive(
       viewModelBuilder: () => CustomizeViewModel(),
+      onModelReady: (CustomizeViewModel model) async => await model.intialize(),
       builder: (
         BuildContext context,
         CustomizeViewModel model,
@@ -70,6 +72,14 @@ class CustomizeView extends StatelessWidget {
                                     ? dangerColor
                                     : lynchColor,
                               ),
+                            ),
+                            style: TextStyle(
+                              fontFamily: 'RobotoMono',
+                              fontSize: blockSize(context) * 1.5,
+                              fontWeight: FontWeight.bold,
+                              color: model.textFieldHasError
+                                  ? dangerColor
+                                  : lynchColor,
                             ),
                             enabled: false,
                           ),
@@ -186,72 +196,112 @@ class CustomizeView extends StatelessWidget {
                                   ],
                                 ),
                                 verticalSpaceSmall(context),
-                                Text(
-                                  'Choose the Flutter channel you want to use:',
-                                  textAlign: TextAlign.start,
+                                model.showAdvanced
+                                    ? Column(
+                                        children: <Widget>[
+                                          Text(
+                                            'Choose the Flutter channel you want to use:',
+                                            textAlign: TextAlign.start,
+                                            style: TextStyle(
+                                              fontFamily: 'RobotoMono',
+                                              fontSize: blockSize(context) * 2,
+                                              fontWeight: FontWeight.bold,
+                                            ),
+                                          ),
+                                          Row(
+                                            children: <Widget>[
+                                              Expanded(
+                                                child: RadioListTile<
+                                                    FlutterChannel>(
+                                                  title: Text(
+                                                    'Stable Channel',
+                                                    textAlign: TextAlign.center,
+                                                    style: TextStyle(
+                                                      fontFamily: 'RobotoMono',
+                                                      fontSize:
+                                                          blockSize(context) *
+                                                              1.5,
+                                                      fontWeight:
+                                                          FontWeight.bold,
+                                                    ),
+                                                  ),
+                                                  value: FlutterChannel.stable,
+                                                  groupValue:
+                                                      model.flutterChannel,
+                                                  onChanged: (FlutterChannel
+                                                      newValue) {
+                                                    model.setFlutterChannel(
+                                                        newValue);
+                                                  },
+                                                ),
+                                              ),
+                                              Expanded(
+                                                child: RadioListTile<
+                                                    FlutterChannel>(
+                                                  title: Text(
+                                                    'Beta Channel',
+                                                    textAlign: TextAlign.center,
+                                                    style: TextStyle(
+                                                      fontFamily: 'RobotoMono',
+                                                      fontSize:
+                                                          blockSize(context) *
+                                                              1.5,
+                                                      fontWeight:
+                                                          FontWeight.bold,
+                                                    ),
+                                                  ),
+                                                  value: FlutterChannel.beta,
+                                                  groupValue:
+                                                      model.flutterChannel,
+                                                  onChanged: (FlutterChannel
+                                                      newValue) {
+                                                    model.setFlutterChannel(
+                                                        newValue);
+                                                  },
+                                                ),
+                                              ),
+                                              Expanded(
+                                                child: RadioListTile<
+                                                    FlutterChannel>(
+                                                  title: Text(
+                                                    'Dev Channel',
+                                                    textAlign: TextAlign.center,
+                                                    style: TextStyle(
+                                                      fontFamily: 'RobotoMono',
+                                                      fontSize:
+                                                          blockSize(context) *
+                                                              1.5,
+                                                      fontWeight:
+                                                          FontWeight.bold,
+                                                    ),
+                                                  ),
+                                                  value: FlutterChannel.dev,
+                                                  groupValue:
+                                                      model.flutterChannel,
+                                                  onChanged: (FlutterChannel
+                                                      newValue) {
+                                                    model.setFlutterChannel(
+                                                        newValue);
+                                                  },
+                                                ),
+                                              ),
+                                            ],
+                                          ),
+                                        ],
+                                      )
+                                    : Container(),
+                                TextLink(
+                                  model.showAdvanced
+                                      ? 'Hide Advanced'
+                                      : 'Show Advanced',
                                   style: TextStyle(
+                                    color: textColorLink,
                                     fontFamily: 'RobotoMono',
-                                    fontSize: blockSize(context) * 2,
-                                    fontWeight: FontWeight.bold,
+                                    fontSize: blockSize(context) * 1.5,
                                   ),
-                                ),
-                                Row(
-                                  children: <Widget>[
-                                    Expanded(
-                                      child: RadioListTile<FlutterChannel>(
-                                        title: Text(
-                                          'Stable Channel',
-                                          textAlign: TextAlign.center,
-                                          style: TextStyle(
-                                            fontFamily: 'RobotoMono',
-                                            fontSize: blockSize(context) * 1.5,
-                                            fontWeight: FontWeight.bold,
-                                          ),
-                                        ),
-                                        value: FlutterChannel.stable,
-                                        groupValue: model.flutterChannel,
-                                        onChanged: (FlutterChannel newValue) {
-                                          model.setFlutterChannel(newValue);
-                                        },
-                                      ),
-                                    ),
-                                    Expanded(
-                                      child: RadioListTile<FlutterChannel>(
-                                        title: Text(
-                                          'Beta Channel',
-                                          textAlign: TextAlign.center,
-                                          style: TextStyle(
-                                            fontFamily: 'RobotoMono',
-                                            fontSize: blockSize(context) * 1.5,
-                                            fontWeight: FontWeight.bold,
-                                          ),
-                                        ),
-                                        value: FlutterChannel.beta,
-                                        groupValue: model.flutterChannel,
-                                        onChanged: (FlutterChannel newValue) {
-                                          model.setFlutterChannel(newValue);
-                                        },
-                                      ),
-                                    ),
-                                    Expanded(
-                                      child: RadioListTile<FlutterChannel>(
-                                        title: Text(
-                                          'Dev Channel',
-                                          textAlign: TextAlign.center,
-                                          style: TextStyle(
-                                            fontFamily: 'RobotoMono',
-                                            fontSize: blockSize(context) * 1.5,
-                                            fontWeight: FontWeight.bold,
-                                          ),
-                                        ),
-                                        value: FlutterChannel.dev,
-                                        groupValue: model.flutterChannel,
-                                        onChanged: (FlutterChannel newValue) {
-                                          model.setFlutterChannel(newValue);
-                                        },
-                                      ),
-                                    ),
-                                  ],
+                                  onPressed: () {
+                                    model.setShowAdvanced(!model.showAdvanced);
+                                  },
                                 ),
                               ],
                             ),