+++ lib/src/ui/views/steps/installing/scripts/linux.dart
@@ -1,0 +1,3 @@
+Future<void> installOnLinux() async {
+  print('Install On Linux');
+}